make changes on files
